refactor(ses): split 7,464-line service.rs into sub-modules
Split monolithic service.rs into 9 files following the SSM pattern:
- mod.rs: struct, dispatch, helpers, tests (3,908 lines)
- misc.rs: tags, dedicated IPs, import/export jobs (1,622 lines)
- configuration_sets.rs: config set CRUD + event destinations (500 lines)
- identities.rs: email identity operations (490 lines)
- contact_lists.rs: contact list + contact CRUD (440 lines)
- templates.rs: email template CRUD (222 lines)
- account.rs: account settings, VDM (173 lines)
- sending.rs: SendEmail, SendBulkEmail (160 lines)
- suppression.rs: suppression list operations (120 lines)

Also updates conformance audit paths for the SES split.
