+++ crates/fakecloud-conformance/src/audit.rs
@@ -20,7 +20,7 @@ fn service_source_files(project_root: &Path) -> Vec<(String, Vec<PathBuf>)> {
         ("logs", "logs", &["service.rs"]),
         ("kms", "kms", &["service.rs"]),
         ("cloudformation", "cloudformation", &["service.rs"]),
-        ("ses", "ses", &["service.rs"]),
+        ("ses", "ses", &["service/mod.rs", "service.rs"]),
         ("cognito-idp", "cognito", &["service.rs"]),
     ];
