+++ crates/fakecloud-ses/src/service/sending.rs
@@ -1,0 +1,160 @@
+use chrono::Utc;
+use http::StatusCode;
+use serde_json::{json, Value};
+
+use fakecloud_core::service::{AwsRequest, AwsResponse, AwsServiceError};
+
+use crate::state::SentEmail;
+
+use super::{extract_string_array, SesV2Service};
+
+impl SesV2Service {
+    pub(super) fn send_email(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
+        let body: Value = Self::parse_body(req)?;
+
+        if !body["Content"].is_object()
+            || (!body["Content"]["Simple"].is_object()
+                && !body["Content"]["Raw"].is_object()
+                && !body["Content"]["Template"].is_object())
+        {
+            return Ok(Self::json_error(
+                StatusCode::BAD_REQUEST,
+                "BadRequestException",
+                "Content is required and must contain Simple, Raw, or Template",
+            ));
+        }
+
+        let from = body["FromEmailAddress"].as_str().unwrap_or("").to_string();
+
+        let to = extract_string_array(&body["Destination"]["ToAddresses"]);
+        let cc = extract_string_array(&body["Destination"]["CcAddresses"]);
+        let bcc = extract_string_array(&body["Destination"]["BccAddresses"]);
+
+        let config_set_name = body["ConfigurationSetName"].as_str().map(|s| s.to_string());
+
+        let (subject, html_body, text_body, raw_data, template_name, template_data) =
+            if body["Content"]["Simple"].is_object() {
+                let simple = &body["Content"]["Simple"];
+                let subject = simple["Subject"]["Data"].as_str().map(|s| s.to_string());
+                let html = simple["Body"]["Html"]["Data"]
+                    .as_str()
+                    .map(|s| s.to_string());
+                let text = simple["Body"]["Text"]["Data"]
+                    .as_str()
+                    .map(|s| s.to_string());
+                (subject, html, text, None, None, None)
+            } else if body["Content"]["Raw"].is_object() {
+                let raw = body["Content"]["Raw"]["Data"]
+                    .as_str()
+                    .map(|s| s.to_string());
+                (None, None, None, raw, None, None)
+            } else if body["Content"]["Template"].is_object() {
+                let tmpl = &body["Content"]["Template"];
+                let tmpl_name = tmpl["TemplateName"].as_str().map(|s| s.to_string());
+                let tmpl_data = tmpl["TemplateData"].as_str().map(|s| s.to_string());
+                (None, None, None, None, tmpl_name, tmpl_data)
+            } else {
+                (None, None, None, None, None, None)
+            };
+
+        let message_id = uuid::Uuid::new_v4().to_string();
+
+        let sent = SentEmail {
+            message_id: message_id.clone(),
+            from,
+            to,
+            cc,
+            bcc,
+            subject,
+            html_body,
+            text_body,
+            raw_data,
+            template_name,
+            template_data,
+            timestamp: Utc::now(),
+        };
+
+        // Event fanout: check suppression list, generate events, deliver to destinations
+        if let Some(ref ctx) = self.delivery_ctx {
+            crate::fanout::process_send_events(ctx, &sent, config_set_name.as_deref());
+        }
+
+        self.state.write().sent_emails.push(sent);
+
+        let response = json!({
+            "MessageId": message_id,
+        });
+
+        Ok(AwsResponse::json(StatusCode::OK, response.to_string()))
+    }
+
+    pub(super) fn send_bulk_email(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
+        let body: Value = Self::parse_body(req)?;
+
+        let from = body["FromEmailAddress"].as_str().unwrap_or("").to_string();
+        let config_set_name = body["ConfigurationSetName"].as_str().map(|s| s.to_string());
+
+        let entries = match body["BulkEmailEntries"].as_array() {
+            Some(arr) if !arr.is_empty() => arr.clone(),
+            _ => {
+                return Ok(Self::json_error(
+                    StatusCode::BAD_REQUEST,
+                    "BadRequestException",
+                    "BulkEmailEntries is required and must not be empty",
+                ));
+            }
+        };
+
+        let mut results = Vec::new();
+
+        for entry in &entries {
+            let to = extract_string_array(&entry["Destination"]["ToAddresses"]);
+            let cc = extract_string_array(&entry["Destination"]["CcAddresses"]);
+            let bcc = extract_string_array(&entry["Destination"]["BccAddresses"]);
+
+            let message_id = uuid::Uuid::new_v4().to_string();
+
+            let template_name = body["DefaultContent"]["Template"]["TemplateName"]
+                .as_str()
+                .map(|s| s.to_string());
+            let template_data = entry["ReplacementEmailContent"]["ReplacementTemplate"]
+                ["ReplacementTemplateData"]
+                .as_str()
+                .or_else(|| body["DefaultContent"]["Template"]["TemplateData"].as_str())
+                .map(|s| s.to_string());
+
+            let sent = SentEmail {
+                message_id: message_id.clone(),
+                from: from.clone(),
+                to,
+                cc,
+                bcc,
+                subject: None,
+                html_body: None,
+                text_body: None,
+                raw_data: None,
+                template_name,
+                template_data,
+                timestamp: Utc::now(),
+            };
+
+            // Event fanout for each bulk entry
+            if let Some(ref ctx) = self.delivery_ctx {
+                crate::fanout::process_send_events(ctx, &sent, config_set_name.as_deref());
+            }
+
+            self.state.write().sent_emails.push(sent);
+
+            results.push(json!({
+                "Status": "SUCCESS",
+                "MessageId": message_id,
+            }));
+        }
+
+        let response = json!({
+            "BulkEmailEntryResults": results,
+        });
+
+        Ok(AwsResponse::json(StatusCode::OK, response.to_string()))
+    }
+}
